% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMBIA.R
\name{createUniquePertbs}
\alias{createUniquePertbs}
\title{Function to make unique perturbations of the replicates these will be used incase if CV is greater than threshold.}
\usage{
createUniquePertbs(totalNumberofReplicates)
}
\arguments{
\item{totalNumberofReplicates}{Total replicate number}
}
\value{
unique possible perturbations
}
\description{
Function to make unique perturbations of the replicates these will be used incase if CV is greater than threshold.
}
\examples{
rslt <- createUniquePertbs(5) 
}
\author{
Muhammad kashif
}
