% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMBIA.R
\name{cVCal}
\alias{cVCal}
\title{This function calculates CV}
\usage{
cVCal(vals)
}
\arguments{
\item{vals}{Values}
}
\value{
cv of input values
}
\description{
This function calculates CV
}
\examples{
mData <- matrix(1:10, 2,5)
rslt <-  cVCal(mData)
}
\author{
Muhammad kashif
}
