% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/COMBIA.R
\name{extractValuesFromRange}
\alias{extractValuesFromRange}
\title{Given a well/wells alphanumeric name e.g. "B2", this function translates it to
numeric values.}
\usage{
extractValuesFromRange(range, excelFormate)
}
\arguments{
\item{range}{Well name e-g B2.}

\item{excelFormate}{TRUE if well name is in spreadsheet format, otherwise cell culture
plate format.}
}
\value{
Index of starting row, ending row, starting column and ending column.
}
\description{
Given a well/wells alphanumeric name e.g. "B2", this function translates it to
numeric values.
}
\examples{
rng <-c("B2")
exclF <- TRUE
rslt <-  extractValuesFromRange(rng, exclF)
}
\author{
Muhammad kashif
}

