% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/COMBIA.R
\name{createUniquePertbs}
\alias{createUniquePertbs}
\title{Function to calculate possible unique  perturbations of replicates.}
\usage{
createUniquePertbs(totalNumberofReplicates)
}
\arguments{
\item{totalNumberofReplicates}{Total replicate number.}
}
\value{
unique possible perturbations.
}
\description{
Function to calculate possible unique  perturbations of replicates.
}
\examples{
rslt<- createUniquePertbs(5)
}
\author{
Muhammad kashif
}

