% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/COMBIA.R
\name{cVCal}
\alias{cVCal}
\title{The function calculates coefficient of variance (CV).}
\usage{
cVCal(vals)
}
\arguments{
\item{vals}{Data.}
}
\value{
CV of input values.
}
\description{
The function calculates coefficient of variance (CV).
}
\examples{
mData<- matrix(1:10, 2,5)
rslt<-  cVCal(mData)
}
\author{
Muhammad kashif
}

