% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMBIA.R
\name{extractValuesFromRange}
\alias{extractValuesFromRange}
\title{This function extract numerical indices of a given range e-g B2}
\usage{
extractValuesFromRange(range, excelFormate)
}
\arguments{
\item{range}{Range e-g B2}

\item{excelFormate}{TRUE if range is in spreadsheet formate}
}
\value{
Number of starting row, ending row, starting column and ending column
}
\description{
This function extract numerical indices of a given range e-g B2
}
\examples{
rng <- c("B2")
exclF <- TRUE
rslt <-  extractValuesFromRange(rng, exclF)
}
\author{
Muhammad kashif
}
